AP_HAL_SITL: sitl system minor fixes fix style fix some cast

+++ libraries/AP_HAL_SITL/SITL_cmdline.cpp
@@ -137,7 +137,7 @@ void SITL_State::_parse_command_line(int argc, char * const argv[])
     _instance = 0;
 
     enum long_options {
-        CMDLINE_CLIENT=0,
+        CMDLINE_CLIENT = 0,
         CMDLINE_GIMBAL,
         CMDLINE_AUTOTESTDIR,
         CMDLINE_UARTA,
@@ -190,13 +190,13 @@ void SITL_State::_parse_command_line(int argc, char * const argv[])
             AP_Param::set_hide_disabled_groups(false);
             break;
         case 'r':
-            _framerate = (unsigned)atoi(gopt.optarg);
+            _framerate = static_cast<uint16_t>(atoi(gopt.optarg));
             break;
         case 'C':
             HALSITL::UARTDriver::_console = true;
             break;
         case 'I': {
-            _instance = atoi(gopt.optarg);
+            _instance = static_cast<uint8_t>(atoi(gopt.optarg));
             _base_port  += _instance * 10;
             _rcout_port += _instance * 10;
             _rcin_port  += _instance * 10;
